split CustomChannelInfo in two parts for future modify endpoint
